(1) Treat floating point consistently as double throughout. (2) Initial rete match/unmatch test passes now.
--HG--
extra : convert_revision : svn%3A849a83f0-50c2-4045-bdd1-8b1f698b4b18/trunk%4013

+++ jsoar/src/org/jsoar/kernel/parser/Lexeme.java
@@ -13,7 +13,7 @@ public class Lexeme
     public String string = "";
     public int type = 0;
     public int int_val;
-    public float float_val;
+    public double float_val;
     public char id_letter;
     public int id_number;
